requirements.txt
langgraph
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
sentence-transformers
pypdf
python-dotenv
requests
beautifulsoup4
lxml
streamlit